refactor: split rest source emitters
