+++ .sampo/changesets/1010-split-rest-source-emitters.md
@@ -1,0 +1,6 @@
+---
+npm/@wp-typia/project-tools: patch
+npm/wp-typia: patch
+---
+
+Split REST workspace source emitters into generated-resource and manual-contract modules while keeping the compatibility facade stable.
